read connections into array via mapfile
When storing the output of nmcli as an array with ($(...)), if a network
name contains whitespace the name itself will be interpreted as a single
array value, without uuid, which breaks the untrusted check. Using
mapfile guarantees that the output is split only at newlines and that
each array value will always be in the "Network Name:UUID" format.