Fix broken status output in ttoggle
The unit name is only shown if the unit is running because ● is used
with grep. Names of failed units are not displayed (they have × in front
of the name).

Unnecessary information is displayed because ● is used with grep:
● roothints.timer - Unbound Root Hints Update Timer
     Active: active (waiting) ...
   Triggers: ● roothints.service

Use sed instead of grep to print the first line with the unit name
together with the status