add support for NetworkManager v1.42
NetworkManager v1.42 introduced loopback interface connections. A
connection is always considered to be active on this interface, and the
UUID changes each time NetworkManager is started (or at least on each
system boot) so the user cannot just mark the loopback interface
connection as trusted. We have to exclude it from any output considered
by nmtrust.

There is probably a less hacky way to do this, but this seems to work.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/blob/main/NEWS